chore: Add API endpoint to get all posts list
This commit adds a new API endpoint `/api/all_posts_list/<current_post_id>` to retrieve the list of all posts. The endpoint returns the current post ID, as well as the IDs of the previous and next posts in the list. This functionality allows for easy navigation between posts.